+++ src/main/java/ZInvest/domain/dto/InntektRegnskapRequest.java
@@ -5,8 +5,8 @@ import java.util.List;
 public class InntektRegnskapRequest {
     private String aar;
     private String label;
-    private Double belop;
-    private List<Double> belopList; //Brukes for å vise inntekt for flere år på regnskap siden
+    private Long belop;
+    private List<Long> belopList; //Brukes for å vise inntekt for flere år på regnskap siden
     private List<String> aarList; //Brukes for å hente detaljer om utgifter på regnskap siden
 
     public String getAar() {
@@ -25,19 +25,19 @@ public class InntektRegnskapRequest {
         this.label = label;
     }
 
-    public Double getBelop() {
+    public Long getBelop() {
         return belop;
     }
 
-    public void setBelop(Double belop) {
+    public void setBelop(Long belop) {
         this.belop = belop;
     }
 
-    public List<Double> getBelopList() {
+    public List<Long> getBelopList() {
         return belopList;
     }
 
-    public void setBelopList(List<Double> belopList) {
+    public void setBelopList(List<Long> belopList) {
         this.belopList = belopList;
     }
 
@@ -60,8 +60,8 @@ public class InntektRegnskapRequest {
     public static class Builder {
         private String aar;
         private String label;
-        private Double belop;
-        private List<Double> belopList;
+        private Long belop;
+        private List<Long> belopList;
         private List<String> aarList;
 
         public Builder() {
@@ -77,12 +77,12 @@ public class InntektRegnskapRequest {
             return this;
         }
 
-        public Builder belop(Double belop) {
+        public Builder belop(Long belop) {
             this.belop = belop;
             return this;
         }
 
-        public Builder belopList(List<Double> belopList) {
+        public Builder belopList(List<Long> belopList) {
             this.belopList = belopList;
             return this;
         }
